[ISSUE-32] move: 임시 이미지 폴더 temp 생성 및 경로 수정 (#26)
